+++ app/src/main/java/de/k3b/android/androFotoFinder/Global.java
@@ -70,8 +70,8 @@ public class Global {
     /** false: cmd setGeo => form(GeoEditActivity) => mapPicker */
     public static boolean geoNoEdit = true;
 
-    /** #26 which image resolution should the non zoomed imageView have? */
-    public static int imageDetailResolution = MediaStore.Images.Thumbnails.MINI_KIND; // FULL_SCREEN_KIND;
+    /** #26 which image resolution should the "non zoomed imageView" have? */
+    public static boolean initialImageDetailResolutionHigh = false; // false: MediaStore.Images.Thumbnails.MINI_KIND; true: FULL_SCREEN_KIND;
 
     public static void debugMemory(String modul, String message) {
         if (Global.debugEnabledMemory) {